requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.9.0
